requirements.txt
apache-airflow
pandas
requests
snowflake-connector-python
boto3
dbt-core
dbt-snowflake
python-dotenv
